setup expenses viewModel

+++ app/src/main/java/com/gdsc/expensetracker/CategoryExpense/RecyclerView/CategoryExpenseRVItem.kt
@@ -2,5 +2,5 @@ package com.gdsc.expensetracker.CategoryExpense.RecyclerView
 
 data class CategoryExpenseRVItem(
     val name : String,
-    val amount : Int
+    val amount : Double
 )
